requirements.txt
flask==3.1.2
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
flask-limiter==3.5.0
psycopg2-binary==2.9.10
tensorflow==2.16.1
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0
werkzeug==3.1.1
requests==2.32.3
pyjwt==2.9.0
python-dotenv==1.0.1 